feat: migrate to modern options format, cleanup
