+++ worlds/simpsonshitnrun/Options.py
@@ -1,35 +1,177 @@
-from Options import FreeText, NumericOption, Toggle, DefaultOnToggle, Choice, TextChoice, Range, NamedRange, PerGameCommonOptions, DeathLink
-from dataclasses import make_dataclass
-from .hooks.Options import before_options_defined, after_options_defined
-from .Data import category_table, game_table
-from .Locations import victory_names
-from .Items import item_table
-
+from dataclasses import dataclass
+from Options import (Toggle, Choice, Range, PerGameCommonOptions, DeathLink)
 
 class FillerTrapPercent(Range):
+    diplay_name = "Filler Traps"
     """How many fillers will be replaced with traps. 0 means no additional traps, 100 means all fillers are traps."""
     range_end = 100
 
-manual_options = before_options_defined({})
 
-if len(victory_names) > 1:
-    goal = {'option_' + v: i for i, v in enumerate(victory_names)}
-    manual_options['goal'] = type('goal', (Choice,), goal)
-    manual_options['goal'].__doc__ = "Choose your victory condition."
+class Goal(Choice):
+    """Choose your victory condition."""
+    display_name = "Goal"
+    default = 0
+    option_goal_all_missions_complete = 0
+    option_goal_all_story_missions_complete = 1
+    option_goal_final_missionl7m7 = 2
+    option_goal_wasps_and_cards_collected = 3
 
-if any(item.get('trap') for item in item_table):
-    manual_options["filler_traps"] = FillerTrapPercent
 
-if game_table.get("death_link"):
-    manual_options["death_link"] = DeathLink
+class LevelSanity(Choice):
+    """Choose how you want missions to be sent in the multiworld. Levels logically
+       require the level access item and at least 1 car from that level or higher.
+       You're guaranteed to start with a level and it's default car with either option.
+       Linear will add progressive levels to the pool. You'll start with Level 1 and
+       the Family Sedan, then get the next level each time you receive a progressive
+       level item. Levels will add levels to the pool. You'll start the game with
+       a random level and it's required items and receive levels from the multiworld.
+       Regardless of your choice, missions can be played in any order on an unlocked level.
+       """
+    display_name = "Levelsanity"
+    option_linear = 0
+    option_levels = 1
+    default = 1
 
-for category in category_table:
-    for option_name in category_table[category].get("yaml_option", []):
-        if option_name[0] == "!":
-            option_name = option_name[1:]
-        if option_name not in manual_options:
-            manual_options[option_name] = type(option_name, (DefaultOnToggle,), {"default": True})
-            manual_options[option_name].__doc__ = "Should items/locations linked to this option be enabled?"
+class MoveRando(Toggle):
+    """Choose whether or not to shuffle moves into the item pool.
+       Moves that are shuffled are Double Jump and Attack for each character.
+       These have logical implications for wasp and card collection
+       as well as L7M4 requiring Homer Double Jump.
+       """
 
-manual_options = after_options_defined(manual_options)
-manual_options_data = make_dataclass('ManualOptionsClass', manual_options.items(), bases=(PerGameCommonOptions,))
+    default = True
+    display_name = "Move Randomizer"
+
+class ShuffleGagfinder(Toggle):
+    """If enabled, add a Gagfinder to the pool for each Character that will be
+       required to unlock gags as that character. If disabled, gags will instead
+       be locked until you receive their Level.
+       """
+
+    default = True
+    display_name = "Shuffle Gagfinder"
+
+class ShuffleCheckeredFlags(Toggle):
+    """If enabled, add a Checkered Flag to the pool for each Character that will be
+       required to unlock races as that character. If disabled, races will instead
+       be locked until you receive their Level.
+       """
+
+    default = True
+    display_name = "Shuffle Checkered Flags"
+
+class ShuffleEBrakes(Toggle):
+    """Choose whether or not to shuffle ability to use the E-Brake
+       for each character into the item pool. *WARNING* This is not
+       considered in logic and has not been heavily tested.
+       It may create unreasonably hard seeds.
+       """
+
+    default = True
+    display_name = "Shuffle E-Brake"
+
+class EnableWaspPercent(Toggle):
+    """Whether to include Wasps in goal requirements.
+       This setting is always treated as true if your
+       goal is Goal: Wasps and Cards Collected!"""
+
+    default = True
+    display_name = "Enable Wasp Requirements"
+
+class WaspPercent(Range):
+    """ Percent of Wasps required to goal if Wasp Requirements is enabled."""
+    display_name = "Required Wasp Percent"
+    range_start = 10
+    range_end = 100
+    default = 50
+
+class EnableCardPercent(Toggle):
+    """Whether to include Cards in goal requirements.
+       This setting is always treated as true if your
+       goal is Goal: Wasps and Cards Collected!"""
+
+    default = True
+    display_name = "Enable Card Requirements"
+
+class CardPercent(Range):
+    """Percent of Cards required to goal if Card Requirements is enabled."""
+    display_name = "Required Card Percent"
+    range_start = 10
+    range_end = 100
+    default = 50
+
+class ShuffleCards(Toggle):
+    """Randomize card locations. This option adds several possible locations for
+       cards. There will still be 49 total cards with 7 in each level.
+       ***THIS OPTION IS UNIMPLEMENTED***"""
+
+    default = True
+    display_name = "Shuffle Cards"
+
+
+class MinShopPrice(Range):
+    """The minimum cost of any item in Gil's Shop. If this is greater than the max shop price, then the max will be used instead."""
+    display_name = "Min Shop Price"
+    range_start = 0
+    range_end = 500
+    default = 100
+
+class MaxShopPrice(Range):
+    """The maximum cost of any item in Gil's Shop."""
+    display_name = "Max Shop Price"
+    range_start = 0
+    range_end = 500
+    default = 300
+
+class ShopScaleMod(Range):
+    """The multiplier for shop costs per levels
+       L2 costs = L1 * multiplier, L3 = L1 * 2(multiplier), etc"""
+    display_name = "Shop Price Level Scale Multiplier"
+    range_start = 1
+    range_end = 5
+    default = 2
+
+class EjectTraps(Toggle):
+    """Whether to include Eject traps in the item pool."""
+    default = True
+    display_name = "Enable Eject Traps"
+
+class DuffTraps(Toggle):
+    """Whether to include Duff traps in the item pool."""
+    default = True
+    display_name = "Enable Duff Traps"
+
+class LaunchTraps(Toggle):
+    """Whether to include Launch traps in the item pool."""
+    default = True
+    display_name = "Enable Launch Traps"
+
+class HNRTraps(Toggle):
+    """Whether to include Hit N Run traps in the item pool."""
+    default = True
+    display_name = "Enable Hit N Run Traps"
+
+@dataclass
+class SimpsonsHitAndRunOptions(PerGameCommonOptions):
+    goal: Goal
+    levelsanity: LevelSanity
+    moverandomizer: MoveRando
+    shufflegagfinder: ShuffleGagfinder
+    shufflecheckeredflags: ShuffleCheckeredFlags
+    shuffleebrake: ShuffleEBrakes
+    EnableWaspPercent: EnableWaspPercent
+    wasppercent: WaspPercent
+    EnableCardPercent: EnableCardPercent
+    cardpercent: CardPercent
+    minprice: MinShopPrice
+    maxprice: MaxShopPrice
+    shopscalemod: ShopScaleMod
+    filler_traps: FillerTrapPercent
+    eject: EjectTraps
+    duff: DuffTraps
+    launch: LaunchTraps
+    hnr: HNRTraps
+    death_link: DeathLink
+
+
+SimpsonsHitAndRunOptions = SimpsonsHitAndRunOptions
